Delvis refaktorering av useFetchRiScs
