+++ plugins/ros/src/components/scenarioWizard/steps/ActionsStep.tsx
@@ -8,7 +8,7 @@ import { TextField } from '../../common/Textfield';
 import AddCircle from '@material-ui/icons/AddCircle';
 import { useFontStyles } from '../../../utils/style';
 import { ActionEdit } from '../components/ActionEdit';
-import { useScenario } from '../../../ScenarioContext';
+import { useScenario } from '../../../contexts/ScenarioContext';
 
 export const ActionsStep = () => {
   const { t } = useTranslationRef(pluginRiScTranslationRef);
